Upgraded to Lucene 7.0.0

+++ core/src/main/java/com/sproutigy/libs/luceneplus/core/search/LuceneSearchResults.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface LuceneSearchResults extends CloseableIterator<LuceneSearchHit> {
     boolean hasTotal();
-    Integer total();
+    Long total();
 
     boolean hasCount();
     Integer count();
